% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_metrics.R
\name{aggregate_metrics}
\alias{aggregate_metrics}
\title{Aggregate metrics and metadata lists}
\usage{
aggregate_metrics(metrics = results, metadata = metas)
}
\arguments{
\item{metrics}{List of metrics dfs}

\item{metadata}{List of metadata dfs}
}
\value{
A list containing 1 aggregated metrics df and 1 aggregasted metadata
df
}
\description{
Takes a list of metrics dfs and a list of metadata dfs,
aggregates them, and returns a list of 1 metric df and 1 metadata df. Also
ensures that value and year columns are coded as character so they are easier
to combine at this stage. Later, they should be turned back to numeric.
}
\keyword{internal}
