% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_census_api.R
\name{call_census_api}
\alias{call_census_api}
\title{Call Census API}
\usage{
call_census_api(
  state_codes,
  years,
  vars,
  census_key,
  region = c("state", "county"),
  sleep_time = 1,
  survey_name = "acs/acs5"
)
}
\arguments{
\item{state_codes}{Character vector of 2-digit state codes (e.g. Vermont =
50)}

\item{years}{Numeric vector of 4-digit years}

\item{vars}{Character vector of census variables}

\item{census_key}{Census API key}

\item{region}{Character string, one of c('county', 'state')}

\item{sleep_time}{Rest between API calls at state/year level (not variable)}

\item{survey_name}{Name of census survey to pull from}
}
\value{
Data frame of results
}
\description{
Download US Census data through API using censusapi package.
Note that instead of conveniently batching variables in calls, we are
calling each iteration of region, year, and variables separately.
Otherwise, if we hit a single error because a variable does not exist in a
certain year, we lose the whole call's worth of data.
}
\keyword{internal}
