% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_usdm_api.R
\name{call_usdm_api}
\alias{call_usdm_api}
\title{Call US Drought Monitor API}
\usage{
call_usdm_api(base, drought_categories, states, start_date, end_date)
}
\arguments{
\item{base}{Base URL as a string, up to but not including the county or state
specification.}

\item{drought_categories}{Vector of integers specifying which drought
severity categories to collect (0 to 4).}

\item{states}{Vector of 2-letter character strings specifying states (e.g.,
VT)}

\item{start_date}{Start date as string in MM/DD/YYYY format, as is accepted
by the API.}

\item{end_date}{End date as string in MM/DD/YYYY format, as is accepted by
the API.}
}
\value{
Loops over states and drought categories to collect list of data,
nested by state and category
}
\description{
Call the US Drought Monitor API. Have to add years manually, map
over them, and set names. Info at
\url{https://droughtmonitor.unl.edu/DmData/DataDownload/WebServiceInfo.aspx#comp}.
}
\keyword{internal}
