% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_n_records.R
\name{check_n_records}
\alias{check_n_records}
\title{Check metric and metadata variable count}
\usage{
check_n_records(
  metric_vars,
  meta_vars,
  section = "Section",
  var_col = "variable_name"
)
}
\arguments{
\item{metric_vars}{A metrics df}

\item{meta_vars}{A metadata df}

\item{section}{Identifier for the script to call out in prints and errors}

\item{var_col}{Name of column that contains variable names}
}
\value{
If pass, prints results to screen. If fail, throw error and identify
the section.
}
\description{
Compares a long-format metrics df and wide-format metadata df to
make sure they contain the same number of variables and that the variables
have the same names. Throws an error if they do not match. Use at the end
of each wrangling script, right after using \code{aggregate_metrics()}, and also
in the aggregation script before exporting.
}
\keyword{internal}
