% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips_key.R
\docType{data}
\name{fips_key}
\alias{fips_key}
\title{FIPS key}
\format{
\subsection{\code{fips_key}}{

A data frame with county and state fips information
\describe{
\item{fips}{
FIPS key. Generally 5 digits: first 2 digits represent state, last 3 digits represent county.
Here, states only have 2 digits (e.g. Vermont is 50), while counties have 5 digits
(e.g. Chittenden County is 50007).
}
\item{county_name}{County name}
\item{state_name}{State name in long format (e.g. Vermont)}
\item{state_code}{
If \code{fips} refers to a state, the 2-character state code is shown here (e.g. VT).
}
}
}
\usage{
fips_key
}
\description{
A \code{df} with FIPS keys for all states and all counties in the Northeast
}
\keyword{datasets}
