% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\docType{data}
\name{keys}
\alias{keys}
\alias{naics_key}
\alias{state_key}
\title{Keys and things}
\format{
Descriptions of NAICS codes
\describe{
\item{naics_key}{something}
\item{state_key}{something}
}
}
\usage{
naics_key
}
\description{
Description here
}
\keyword{datasets}
