% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_wider.R
\name{make_wider}
\alias{make_wider}
\title{Make metrics df wider}
\usage{
make_wider(df, var_col = "variable_name", year_col = "year", val_col = "value")
}
\arguments{
\item{df}{\code{metrics} df in long format}

\item{var_col}{column name representing variable name}

\item{year_col}{column name representing year}

\item{val_col}{column name representing value}
}
\value{
\code{metrics} df in wide format
}
\description{
Pivots wider based on defaults used in metrics data. Automatically pastes the
year to the name of new variable columns (\code{foodInsecurity} in 2023 becomes
\code{foodInsecurity_2023}).
}
\examples{
metrics <- data.frame(
  fips = c(50001, 50002, 50003),
  year = 2023,
  variable_name = 'var',
  value = c(1, 2, 3)
)
make_wider(metrics)
}
