% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_citation.R
\name{meta_citation}
\alias{meta_citation}
\title{Metadata: Add citation column}
\usage{
meta_citation(
  df,
  date = format(Sys.Date(), "\%B \%d, \%Y"),
  source_col = source,
  url_col = url,
  api_url = NULL
)
}
\arguments{
\item{df}{A metadata df}

\item{date}{Date in long format (e.g. June 15, 2025). If none given, will
default to the current day.}

\item{source_col}{Name of column that contains the source without quotes.}

\item{url_col}{Name of column with URL without quotes.}

\item{api_url}{If the API URL is different from the listed source, add it
here and it will be appropriated formatted.}
}
\value{
The metadata df with a new citation column
}
\description{
Function to be used in metadata pipeline after metadata df is
created. Pipe it through to add a new citation column which builds on the
source column and either the URL column, or, if the user adds an api_url,
it builds on that.
}
\details{
Note that we need to add an option for the access date, because if
it is a journal or something, we shouldn't have a date retrieved. But most
data are coming from websites or APIs anyway.
}
\keyword{internal}
