% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_latest_year.R
\name{meta_latest_year}
\alias{meta_latest_year}
\title{Metadata: Get latest year per variable}
\usage{
meta_latest_year(df)
}
\arguments{
\item{df}{A long-format metrics df.}
}
\value{
Vector of years, one for each variable, in order of variable names
}
\description{
Get a vector of latest years, one for each variable_name from a metrics-style
long-format df. To be used in metadata creation.
}
