% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_resolution.R
\name{meta_resolution}
\alias{meta_resolution}
\title{Metadata: Get resolution per variable}
\usage{
meta_resolution(df)
}
\arguments{
\item{df}{A long-format metrics df.}
}
\value{
Vector of resolutions ("state", "county", "county, state"), one for
each variable, in order of variable names.
}
\description{
Get a vector of resolution, one for each variable_name from a
metrics-style long-format df. To be used in metadata creation.
}
