% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_vars.R
\name{meta_vars}
\alias{meta_vars}
\title{Metadata: Get variable names}
\usage{
meta_vars(df, col_name = "variable_name")
}
\arguments{
\item{df}{A long-format metrics df.}

\item{col_name}{Name of the column that contains the variable_name}
}
\value{
Sorted vector of unique variable_name
}
\description{
Get a sorted vector of variable names from a metrics-style long-format df. To
be used in metadata creation.
}
