% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_years.R
\name{meta_years}
\alias{meta_years}
\title{Metadata: Get all years per variable}
\usage{
meta_years(df)
}
\arguments{
\item{df}{A long-format metrics df.}
}
\value{
Sorted vector of years for each variable_name. Variable_name is also
returned in sorted order.
}
\description{
Get a sorted vector of years for each variable_name from a metrics-style
long-format df. To be used in metadata creation.
}
