% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_utilities.R
\name{get_vars}
\alias{get_vars}
\alias{get_all_years}
\alias{get_max_year}
\alias{get_resolution}
\alias{add_citation}
\alias{aggregate_metrics}
\alias{check_n_records}
\title{Metadata Utilities}
\usage{
get_vars(df, col_name = "variable_name")

get_all_years(df)

get_max_year(df)

get_resolution(df)

add_citation(
  df,
  date = format(Sys.Date(), "\%B \%d, \%Y"),
  source_col = source,
  url_col = url,
  api_url = NULL
)

aggregate_metrics(metrics = results, metadata = metas)

check_n_records(
  metric_vars,
  meta_vars,
  section = "Section",
  var_col = "variable_name"
)
}
\description{
Functions to build metadata
\itemize{
\item \code{get_vars()}: Input a long format metrics df and return a vector of unique,
sorted variable names.
\item \code{get_all_years()}: Input a long format metrics df and return a character
string of all unique years to put in metadata. Return a vector of strings if
there is more than one unique variable.
\item \code{get_max_year()}: Input a long format metrics df and return a character
string with the latest year represented. Returns a vector of strings if
there is more than one unique variable.
\item \code{get_resolution()}: Input a long format metrics df and return a character
string or vector of strings describing the resolution for each metric as
either county or state. Determined based on the format of the FIPS keys
for each variable.
\item \code{add_citation()}: Input a metadata df and return the df with an extra
column that combines the source with either the URL or the API URL. If API,
adds the access date to the citation.
\item \code{aggregate_metrics()}: Input a list of metrics dfs and a list of metadata
dfs and return a list of two aggregated dfs, one for metrics and one for
metadata. Use at end of wrangling script before \code{check_n_records()}.
\item \code{check_n_records()}: Input a metric df and metadata df along with the
section or script the function appears in and function will check to see
whether all the metrics in the metric df match those in the metadata df. If
they are not the same names and same counts, throws an error. Use at end of
each wrangling script.
}
}
\keyword{internal}
