% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_example.R
\docType{data}
\name{metrics_example}
\alias{metrics_example}
\title{Metrics EXAMPLE}
\format{
\subsection{\code{metrics_example}}{

A data frame in long format with one value per row.
\describe{
\item{fips}{
Federal Information Processing Standards code, a 5-digit code identifying
the location. The first two digits are state, last three are county. Here,
state data is given only 2 digits (e.g. Vermont is \code{50}), while counties
are given all 5 (e.g. Chittenden county is \code{50007}).
}
\item{year}{Year in \code{YYYY} format.}
\item{variable_name}{
String used to identify metric and link it to metadata. Meant for internal
use. Human-friendly metric names are found in the metdata.
}
\item{value}{Value of metric. See metadata for units and interpretation.}
}
}
}
\usage{
metrics_example
}
\description{
Just a small example of metrics dataset to use in testing and function
examples. DO NOT USE IN ANALYSES.
}
\keyword{datasets}
