% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_variable.R
\name{pull_variable}
\alias{pull_variable}
\title{Pull NASS Variable}
\usage{
pull_variable(
  df,
  sector_desc = NULL,
  commodity_desc = NULL,
  statisticcat_desc = NULL,
  domain_desc = NULL,
  short_desc = NULL,
  variable_name = NULL,
  source_desc = "CENSUS"
)
}
\arguments{
\item{df}{}

\item{sector_desc}{NASS sector description}

\item{commodity_desc}{NASS commodity description}

\item{statisticcat_desc}{NASS statistic category description}

\item{domain_desc}{NASS domain description}

\item{short_desc}{NASS short variable description}

\item{variable_name}{short, unique, camel-case identifier for variable}

\item{source_desc}{NASS source description}
}
\description{
Internal function to pull a particular variable from a bulk API download from NASS data. Only
relevant for the \code{nass_compilation.R} script. Function will filter it, name it, and remove NAs
}
\examples{
\dontrun{
results$total_income <- pull_variable(
  bulk_nass_df,
  sector_desc = 'ECONOMICS',
  commodity_desc = 'INCOME, FARM-RELATED',
  domain_desc = 'TOTAL',
  short_desc = 'INCOME, FARM-RELATED - RECEIPTS, MEASURED IN $ / OPERATION',
  variable_name = 'farmIncomePF'
)
}
}
\keyword{internal}
