% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_all_rds.R
\name{read_all_rds}
\alias{read_all_rds}
\title{Read all RDS files at a given location}
\usage{
read_all_rds(path, pattern = NULL, split = "\\\\.")
}
\arguments{
\item{path}{File path from which to load RDS files}

\item{pattern}{regex matching pattern}

\item{split}{What character to split on to separate file suffix from name}
}
\value{
List of RDS files
}
\description{
For internal use only, really just in the aggregation and export stages.
}
