% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_call_apis.R
\name{sm_call_apis}
\alias{sm_call_apis}
\title{SM Call APIs}
\usage{
sm_call_apis(
  nass = FALSE,
  bls = FALSE,
  census = FALSE,
  fda = FALSE,
  usdm = FALSE,
  all = FALSE
)
}
\arguments{
\item{nass}{}

\item{bls}{}

\item{census}{}

\item{fda}{}

\item{usdm}{}

\item{all}{If \code{TRUE}, run all API scripts.}
}
\value{
Runs all APIs and outputs are saved to "5_outputs/api_outs/"
}
\description{
Call all APIs used in secondary data collection. Note that this
function will take some time to run. API keys are invoked from .Renviron file
which is called in the appropriate scripts.
}
\keyword{internal}
