% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_export.R
\name{sm_export}
\alias{sm_export}
\title{SM Export}
\usage{
sm_export(update = TRUE, load_all = TRUE)
}
\arguments{
\item{update}{If \code{TRUE}, run \code{sm_update()} to update the \code{.rda} files in the
project.}
}
\value{
Output data to "6_outputs/". "sm_data.rds" contains metrics, keys,
and probably other things. "sm_spatial.rds" contains state and county
polygons as well as a few other choice spatial datasets. The function also
creates "metadata.csv" and "metrics.csv" which can be used to pick up
analysis in other languages if necessary.
}
\description{
Package and export the metrics, metadata, and utility objects.
Noting that the aggregate and export scripts should probably be combined
into one at some point.
}
\keyword{internal}
