% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_update.R
\name{sm_update}
\alias{sm_update}
\title{Update exported datasets}
\usage{
sm_update()
}
\description{
Internal function that updates the .rda files exported by the package. This
should come after using the \code{sm_aggregate()} and \code{sm_export()} functions
which update the .rds files in the outputs folder.
}
\keyword{internal}
