% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_wrangle.R
\name{sm_wrangle}
\alias{sm_wrangle}
\title{SM Wrangle}
\usage{
sm_wrangle(spatial = FALSE)
}
\arguments{
\item{spatial}{If \code{TRUE}, run the spatial wrangling script, which will take
at least 15 minutes or so to run.}
}
\value{
Runs all data wrangling scripts. Each one saves metric outputs to
"5_objects/results" and metadata to "5_objects/metadata".
}
\description{
SM Wrangle
}
\keyword{internal}
