% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_key.R
\docType{data}
\name{state_key}
\alias{state_key}
\title{State Key}
\format{
\subsection{\code{state_key}}{

A data frame with county and state fips information
\describe{
\item{state}{2-digit state name}
\item{state_code}{2-digit fips code}
\item{state_name}{Full state name}
\item{full_state_code}{5-digit FIPS code (e.g. Vermont is 50000)}
}
}
}
\usage{
state_key
}
\description{
All states in US, including names, abbreviated names, FIPS codes, and abbreviated FIPS codes
}
\keyword{datasets}
