% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_census_api.R
\name{call_census_api}
\alias{call_census_api}
\title{Call Census API}
\usage{
call_census_api(survey_year, survey, vars, county = "*", state)
}
\arguments{
\item{survey_year}{Year as integer. Note that ACS5 is conducted on 5-year
cycle, last one in 2022.}

\item{survey}{String describing which Census survey to pull from (e.g.,
'acs/acs5')}

\item{vars}{String containing one or more variable codes separated by commas
(e.g., "B01003_001E,B15003_001R")}

\item{county}{String containing one or more county FIPS codes separated by
commas (e.g., "50001,50002")}

\item{state}{String containing one or more state FIPS codes separated by
commas (e.g., "50,25")}
}
\value{
List of API outputs for the given year
}
\description{
Call Census API by year. For multiple years, loop the function
over a vector of years. Note that we still need to add a Sys.time() so that
we don't get rate limited. Also, consider the \code{censusapi} package, which
might have been a better way to do this from the beginning.
}
\keyword{internal}
