% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_utilities.R
\name{aggregate_metrics}
\alias{aggregate_metrics}
\alias{check_n_records}
\title{Metadata Utilities}
\usage{
aggregate_metrics(metrics = results, metadata = metas)

check_n_records(
  metric_vars,
  meta_vars,
  section = "Section",
  var_col = "variable_name"
)
}
\description{
Functions to build metadata
\itemize{
\item \code{aggregate_metrics()}: Input a list of metrics dfs and a list of metadata
dfs and return a list of two aggregated dfs, one for metrics and one for
metadata. Use at end of wrangling script before \code{check_n_records()}.
\item \code{check_n_records()}: Input a metric df and metadata df along with the
section or script the function appears in and function will check to see
whether all the metrics in the metric df match those in the metadata df. If
they are not the same names and same counts, throws an error. Use at end of
each wrangling script.
}
}
\keyword{internal}
