% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_update.R
\name{sm_update}
\alias{sm_update}
\title{Update exported datasets}
\usage{
sm_update(export = FALSE)
}
\arguments{
\item{export}{If TRUE, run \code{sm_export()} first.}
}
\description{
Internal function that updates the .rda files exported by the package. This
should come after using the \code{sm_export()} function which updates the .rds
files in the outputs folder. Updates only the essentials from the spatial
data, including counties, states, and the mask for the Northeast region.
}
\keyword{internal}
